# postman/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vouchers": {
      "post": {
        "operationId": "VouchersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoucherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a new voucher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoucherEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new voucher",
        "tags": [
          "Vouchers"
        ]
      },
      "get": {
        "operationId": "VouchersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all vouchers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoucherEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all vouchers",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/vouchers/{id}": {
      "get": {
        "operationId": "VouchersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a voucher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoucherEntity"
                }
              }
            }
          }
        },
        "summary": "Get a voucher",
        "tags": [
          "Vouchers"
        ]
      },
      "patch": {
        "operationId": "VouchersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVoucherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a voucher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoucherEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a voucher",
        "tags": [
          "Vouchers"
        ]
      },
      "delete": {
        "operationId": "VouchersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a voucher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoucherEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a voucher",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/vouchers/{id}/all-categories-applied": {
      "get": {
        "operationId": "VouchersController_getAllCategoriesAreAppliedThisVoucher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all categories are applied this voucher"
          }
        },
        "summary": "Get all categories are applied this voucher",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/vouchers/{id}/all-products-applied": {
      "get": {
        "operationId": "VouchersController_getAllProductsAreAppliedThisVoucher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all products are applied this voucher"
          }
        },
        "summary": "Get all products are applied this voucher",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/vouchers/{id}/all-product-variants-applied": {
      "get": {
        "operationId": "VouchersController_getAllProductVariantsAreAppliedThisVoucher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all product-variants are applied this voucher"
          }
        },
        "summary": "Get all product-variants are applied this voucher",
        "tags": [
          "Vouchers"
        ]
      }
    },
...
  },
  "info": {
    "title": "BE Ecomerce Shop",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "VoucherEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "code": {
            "type": "string",
            "example": "SAVE20"
          },
          "description": {
            "type": "object",
            "example": "20% off on all items",
            "nullable": true
          },
          "discountType": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT"
            ],
            "example": "FIXED_AMOUNT"
          },
          "discountValue": {
            "type": "number",
            "example": 20
          },
          "validFrom": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "validTo": {
            "format": "date-time",
            "type": "string",
            "example": "2024-12-31T23:59:59.999Z"
          },
          "usageLimit": {
            "type": "object",
            "example": 100,
            "nullable": true
          },
          "timesUsed": {
            "type": "number",
            "example": 5
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdBy": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "code",
          "description",
          "discountType",
          "discountValue",
          "validFrom",
          "validTo",
          "usageLimit",
          "timesUsed",
          "isActive",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateVoucherDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "fdsakjfhds4655456ewrew"
          },
          "description": {
            "type": "string",
            "example": "free ship"
          },
          "discountType": {
            "type": "string",
            "example": "FIXED_AMOUNT"
          },
          "discountValue": {
            "type": "number",
            "example": 465
          },
          "validFrom": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "validTo": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "usageLimit": {
            "type": "number",
            "example": 4741
          },
          "timesUsed": {
            "type": "number",
            "example": 4741
          },
          "isActive": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "createdBy": {
            "format": "int64",
            "type": "integer",
            "example": 4567
          }
        },
        "required": [
          "code",
          "description",
          "discountType",
          "discountValue",
          "validFrom",
          "validTo",
          "usageLimit",
          "timesUsed",
          "isActive",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
      "UpdateVoucherDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "fdsakjfhds4655456ewrew"
          },
          "description": {
            "type": "string",
            "example": "free ship"
          },
          "discountType": {
            "type": "string",
            "example": "FIXED_AMOUNT"
          },
          "discountValue": {
            "type": "number",
            "example": 465
          },
          "validFrom": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "validTo": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "usageLimit": {
            "type": "number",
            "example": 4741
          },
          "timesUsed": {
            "type": "number",
            "example": 4741
          },
          "isActive": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:21.006Z"
          },
          "createdBy": {
            "format": "int64",
            "type": "integer",
            "example": 4567
          }
        }
      },
...
    }
  }
}